requirements.txt
# Core
python-dotenv
ipykernel
pandas
numexpr
validators

# LangChain ecosystem
langchain
langchain-community
langchain-text-splitters
langchain-openai
langchain-huggingface
langchain-groq
langchain-chroma
langchain_nvidia_ai_endpoints

# LLM / Embeddings
sentence-transformers
torch
faiss-cpu
chromadb

# Data loaders
pypdf
pymupdf
bs4
wikipedia
arxiv
duckduckgo_search
youtube_transcript_api
unstructured[pdf]
pytube

# DB
duckdb
mysql-connector-python
SQLAlchemy

# Hugging Face
huggingface_hub

# Streamlit
streamlit
